+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## unreleased
+## 0.5.0
 
 - Update to `symbolite >= 0.6` to support serialization (pickle) of `System`.
 - Add `__getnewargs__` to `Variable` and `Derivative` to support serialization